--- src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/deliveryRule/dto/DeliveryRuleGetResponse.java
@@ -1,20 +1,0 @@
-package com.nhnacademy.mini_dooray.ssacthree_front.admin.deliveryRule.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class DeliveryRuleGetResponse {
-
-    private long deliveryRuleId;
-    private String deliveryRuleName;
-    private int deliveryFee;
-    private int deliveryDiscountCost;
-    private boolean deliveryRuleIsSelected;
-    private LocalDateTime deliveryRuleCreatedAt;
-}
