--- src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/deliveryRule/exception/DeliveryRuleCreateFailedException.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.mini_dooray.ssacthree_front.admin.deliveryRule.exception;
-
-public class DeliveryRuleCreateFailedException extends RuntimeException {
-    public DeliveryRuleCreateFailedException(String message) {
-        super(message);
-    }
-}
