+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/delivery_rule/dto/DeliveryRuleGetResponse.java
@@ -1,0 +1,20 @@
+package com.nhnacademy.mini_dooray.ssacthree_front.admin.delivery_rule.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class DeliveryRuleGetResponse {
+
+    private long deliveryRuleId;
+    private String deliveryRuleName;
+    private int deliveryFee;
+    private int deliveryDiscountCost;
+    private boolean deliveryRuleIsSelected;
+    private LocalDateTime deliveryRuleCreatedAt;
+}
