+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/delivery_rule/exception/DeliveryRuleCreateFailedException.java
@@ -1,0 +1,7 @@
+package com.nhnacademy.mini_dooray.ssacthree_front.admin.delivery_rule.exception;
+
+public class DeliveryRuleCreateFailedException extends RuntimeException {
+    public DeliveryRuleCreateFailedException(String message) {
+        super(message);
+    }
+}
