--- src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/pointSaveRule/domain/PointSaveType.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.mini_dooray.ssacthree_front.admin.pointSaveRule.domain;
-
-public enum PointSaveType {
-    PERCENT,INTEGER
-}
