--- src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/pointSaveRule/exception/PointSaveRuleCreateFailedException.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.mini_dooray.ssacthree_front.admin.pointSaveRule.exception;
-
-public class PointSaveRuleCreateFailedException extends RuntimeException {
-    public PointSaveRuleCreateFailedException(String message) {
-        super(message);
-    }
-}
