+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/point_save_rule/domain/PointSaveType.java
@@ -1,0 +1,5 @@
+package com.nhnacademy.mini_dooray.ssacthree_front.admin.point_save_rule.domain;
+
+public enum PointSaveType {
+    PERCENT,INTEGER
+}
