+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/point_save_rule/dto/PointSaveRuleUpdateRequest.java
@@ -1,0 +1,17 @@
+package com.nhnacademy.mini_dooray.ssacthree_front.admin.point_save_rule.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class PointSaveRuleUpdateRequest {
+
+    @NotNull
+    private Long pointSaveRuleId;
+}
