+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/admin/point_save_rule/exception/PointSaveRuleCreateFailedException.java
@@ -1,0 +1,7 @@
+package com.nhnacademy.mini_dooray.ssacthree_front.admin.point_save_rule.exception;
+
+public class PointSaveRuleCreateFailedException extends RuntimeException {
+    public PointSaveRuleCreateFailedException(String message) {
+        super(message);
+    }
+}
