+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/bookset/category/dto/response/CategoryInfoResponse.java
@@ -1,9 +1,5 @@
 package com.nhnacademy.mini_dooray.ssacthree_front.bookset.category.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
