+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/member/dto/AddressRequest.java
@@ -1,13 +1,10 @@
 package com.nhnacademy.mini_dooray.ssacthree_front.member.dto;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Null;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @NoArgsConstructor
 @AllArgsConstructor
