+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/member/dto/MemberInfoResponse.java
@@ -4,7 +4,6 @@ package com.nhnacademy.mini_dooray.ssacthree_front.member.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @AllArgsConstructor
