+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/member/exception/CustomerNotFoundException.java
@@ -1,0 +1,8 @@
+package com.nhnacademy.mini_dooray.ssacthree_front.member.exception;
+
+public class CustomerNotFoundException extends RuntimeException {
+
+    public CustomerNotFoundException(String message) {
+        super(message);
+    }
+}
