+++ src/main/java/com/nhnacademy/mini_dooray/ssacthree_front/member/service/AddressService.java
@@ -1,6 +1,5 @@
 package com.nhnacademy.mini_dooray.ssacthree_front.member.service;
 
-import com.nhnacademy.mini_dooray.ssacthree_front.commons.dto.MessageResponse;
 import com.nhnacademy.mini_dooray.ssacthree_front.member.dto.AddressRequest;
 import com.nhnacademy.mini_dooray.ssacthree_front.member.dto.AddressResponse;
 import jakarta.servlet.http.HttpServletRequest;
